move adj-list-map to own file
