+++ algorithms/src/graph-kosaraju.js
@@ -1,5 +1,7 @@
 // @ts-check
 
+const { Graph } = require('./graph-adj-list-map');
+
 /**
  * Kosaraju's Strongly Connected Component algorithm
  * @param {Graph} g
@@ -19,63 +21,6 @@ function kosaraju(g) {
 }
 
 ///////////////////////////////////////////////////////////////
-
-class Graph {
-	/**
-	 * Constructs a directed graph using an adjacency list.
-	 */
-	constructor() {
-		/** @type {Map<string, string>} */
-		this.nodes = new Map();
-
-		/** @type {Map<string, []>} */
-		this.edges = new Map();
-	}
-
-	/**
-	 * Adds a node to the list in O(1)
-	 * @param {string} val
-	 */
-	addNode(val) {
-		this.nodes.set(val, val);
-		this.edges.set(val, []);
-	}
-
-	/**
-	 * Adds a directed edge to the list in O(1)
-	 * @param {string} tail
-	 * @param {string} head
-	 */
-	addEdge(tail, head) {
-		let edges = this.getEdges(tail);
-		edges.push(head);
-	}
-
-	/**
-	 * Gets a node by the value specified
-	 * @param {string} val
-	 */
-	getNode(val) {
-		return this.nodes.get(val);
-	}
-
-	/**
-	 * Gets all nodes
-	 * @returns {string[]}
-	 */
-	getNodes() {
-		return Array.from(this.nodes.values());
-	}
-
-	/**
-	 * Gets the outbound edges for a node
-	 * @param {string} val
-	 * @returns {string[]}
-	 */
-	getEdges(val) {
-		return this.edges.get(val);
-	}
-}
 
 /**
  * Reverses the graph by constructing a duplicate of it with all
@@ -229,7 +174,6 @@ function findLeaders(g, order) {
 	return leader;
 }
 
-exports.Graph = Graph;
 exports.reverseGraph = reverseGraph;
 exports.findFinishingOrder = findFinishingOrder;
 exports.findLeaders = findLeaders;
